Do not reuse the same socket for multiple requests
It has been reported that sometimes the alarm system replies are not
what we expected them to be.

To trigger this bug you need to create an IAlarm instance and call
"get_status()" in loop. Then with the official iAlarm application change
the alarm status from disarmed to armed.
At this point "get_status()" will start to complain that it doesn't
understand the alarm system replies.

By using a fresh new socket connection it seems like to be way safer and
this bug doesn't happen to present itself anymore.