Use again the sequence number when sending request list
I stopped using the sequence number because I was resetting the
connection every time and it seemed like it had no effect.

But after receiving a report that after the latest update sometimes the
request list reply is empty, it's better to start reusing the sequence
number. The worst case scenario is that this has no effect, so it's
better to try in keeping it.