Deprecate get_status memory parameter
The only way to get the "triggered" status is by checking the sensors
status.
When not using the memory feature, we were always skipping the sensors
status check, and that leaded to potentially missing reporting the
triggered status.

Now `include_memory_feature` had been deprecated and `get_status()` will
always act as it the memory feature was set to true.

Fixes: https://github.com/RyuzakiKK/pyialarm/issues/9